+++ service/calc/add.go
@@ -1,9 +1,9 @@
 package calc
 
 // Add method adds an array of int64
-func Add(args []int64) (result float64) {
+func Add(args []int64) (result float32) {
 	for _, s := range args {
-		result += float64(s)
+		result += float32(s)
 	}
 	return
 }
