+++ service/calc/sub_test.go
@@ -9,7 +9,7 @@ func TestSub(t *testing.T) {
 	tests := []struct {
 		name       string
 		args       args
-		wantResult float64
+		wantResult float32
 	}{
 		{
 			name:       "Able to execute happy path",
